Improve fixture:stat:prepare task

+++ spec/support/fixture_stat_task.sh
@@ -1,5 +1,15 @@
 #shellcheck shell=sh
 
+# without root privileges
+# $ shellspec --task fixture:stat:prepare
+#
+# with root privileges
+# $ sudo $(which shellspec) --task fixture:stat:prepare
+#
+# cleanup
+# $ shellspec --task fixture:stat:cleanup
+
+
 set -eu
 
 task "fixture:stat:prepare" "Prepare file stat tests"
@@ -8,28 +18,73 @@ task "fixture:stat:cleanup" "Cleanup file stat tests"
 fixture="$SHELLSPEC_SPECDIR/fixture"
 owner=${SUDO_UID:-$(id -u)}:${SUDO_GID:-$(id -g)}
 
-symlink() { ln -s ../file "$1" && chown -h "$owner" "$1"; }
-pipe() { mkfifo "$1" && chown "$owner" "$1"; }
-socket() { nc -lU "$1" & sleep 1 && kill $! && chown "$owner" "$1"; }
-file() { touch "$1" && chown "$owner" "$1" && chmod "$2" "$1"; }
-device() { mknod "$@" && chown "$owner" "$1"; }
+symlink() {
+  ln -s ../file "$1" 2>/dev/null
+  if [ -L "$1" ]; then
+    chown -h "$owner" "$1"
+    return 0
+  fi
+  rm "$1"
+  return 1
+}
+
+pipe() {
+  mkfifo "$1" 2>/dev/null
+  if [ -p "$1" ]; then
+    chown "$owner" "$1"
+    return 0
+  fi
+  rm "$1"
+  return 1
+}
+
+socket() {
+  if nc -lU "$1" 2>/dev/null & then
+    sleep 1
+    kill $!
+    chown "$owner" "$1"
+    return 0
+  fi
+  return 1
+}
+
+file() {
+  echo "${4:-}" > "$1"
+  chown "$owner" "$1"
+  chmod "$2" "$1"
+  case $(ls -dl "$1") in
+    $3*) return 0
+  esac
+  rm "$1"
+  return 1
+}
+
+device() {
+  mknod "$@" 2>/dev/null
+  if [ -e "$1" ]; then
+    chown "$owner" "$1" 2>/dev/null && return 0
+    rm -f "$1"
+  fi
+  return 1
+}
 
 create() {
-  if [ -e "$2" ]; then
-    echo "exist '$2'"
-  elif "$@"; then
-    echo "created '$2'"
+  [ -e "$2" ] && rm -f "$2"
+
+  if "$@"; then
+    echo "[created] '$2'"
   else
-    echo "can not create '$2'"
+    echo "[failure] '$2'"
   fi
 }
 
 delete() {
   if [ -e "$2" ]; then
-    rm -f "$2"
-    echo "deleted '$2'"
-  else
-    echo "not exist '$2'"
+    if rm -f "$2"; then
+      echo "[deleted] '$2'"
+    else
+    echo "[failure] '$2'"
+    fi
   fi
 }
 
@@ -37,12 +92,12 @@ fixture_stat_files() {
   "$1" symlink "$fixture/stat/symlink"
   "$1" pipe    "$fixture/stat/pipe"
   "$1" socket  "$fixture/stat/socket"
-  "$1" file    "$fixture/stat/readable"         a=,u+r
-  "$1" file    "$fixture/stat/writable"         a=,u+w
-  "$1" file    "$fixture/stat/executable"       a=,u+x
-  "$1" file    "$fixture/stat/no-permission"    a=
-  "$1" file    "$fixture/stat/setgid"           a=,g+s
-  "$1" file    "$fixture/stat/setuid"           a=,u+s
+  "$1" file    "$fixture/stat/readable"         a=,u+r "-r??"
+  "$1" file    "$fixture/stat/writable"         a=,u+w "-?w?"
+  "$1" file    "$fixture/stat/executable"       a=,u+x "-??x"     "#!/bin/sh"
+  "$1" file    "$fixture/stat/no-permission"    a=     "----"
+  "$1" file    "$fixture/stat/setuid"           a=,u+s "---S"
+  "$1" file    "$fixture/stat/setgid"           a=,g+s "------S"
   "$1" device  "$fixture/stat/block-device"     b 0 0 # Unnamed devices
   "$1" device  "$fixture/stat/charactor-device" c 1 3 # Null device
 }
